+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [Unreleased]
 
 ### Added
+- Add Test button to ExerciseEditorPage (#131)
 
 ### Fixed
 - Fix toast to be proper floating notification with fixed positioning (#106)